LayoutTests/platform/gtk/inspector/css/shadow-scoped-style-expected.txt

Test matched styles for nodes include shadow dom scoped styles.

Before
After

== Running test suite: CSS.ShadowScopedStyle
-- Running test case: CSS.ShadowScopedStyle.HostStyles
PASS: Node should have shadow roots.
PASS: Node should not be in a shadow tree.
PASS: Node should have an open shadow root.
PASS: Node should have no pseudo element styles.
STYLES:
#shadow-host {
    padding: 5px;
}
#shadow-host {
    padding: 10px;
}
:host {
    padding: 20px;
}
address, article, aside, div, footer, header, hgroup, layer, main, nav, section {
}
html {
}

-- Running test case: CSS.ShadowScopedStyle.ElementInShadowTree
PASS: Node should not have shadow roots.
PASS: Node should be in a shadow tree.
STYLES:
#shadow-node {
    color: green;
}
div {
    color: blue;
}
address, article, aside, div, footer, header, hgroup, layer, main, nav, section {
}
